+++ crates/matrix-sdk-crypto/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Changes:
 
+- Add message IDs to all outgoing to-device messages encrypted by
+  `matrix-sdk-crypto`. The `message-ids` feature of `matrix-sdk-crypto` and
+  `matrix-sdk-base` is now a no-op.
+  ([#3776](https://github.com/matrix-org/matrix-rust-sdk/pull/3776))
+
 - Log the content of received `m.room_key.withheld` to-device events.
   ([#3591](https://github.com/matrix-org/matrix-rust-sdk/pull/3591))